+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [Optimize: rename opertion EQUAL to EQUALS](https://github.com/Tencent/spring-cloud-tencent/pull/457)
 - [Docs: Add featureenv example readme(en) file](https://github.com/Tencent/spring-cloud-tencent/pull/460)
 - [docs:update configuration metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/463)
+- [feat:rename the configuration of reporter.](https://github.com/Tencent/spring-cloud-tencent/pull/465)